DBZ-8924 Move `TransactionCommitConsumer`
